+++ lib/core.dart
@@ -9,6 +9,7 @@ export 'src/core/pair.dart';
 export 'src/core/req.dart';
 export 'src/core/req_digit.dart';
 export 'src/core/req_len.dart';
+export 'src/core/req_num.dart';
 export 'src/core/rules.dart';
 export 'src/core/to_val_obj.dart';
 export 'src/core/trim.dart';