+++ lib/src/core/num.dart
@@ -1,6 +1,6 @@
 import 'package:formdator/formdator.dart';
 
-/// Numeric constraints — limits input to numeric values.
+/// Numeric constraints — limits input data to numeric values.
 ///
 /// Examples:
 /// - valid input: "1", "123", "-1", "-9.99"
@@ -8,52 +8,58 @@ import 'package:formdator/formdator.dart';
 class Num {
   /// Constrains data to numeric values.
   ///
-  /// [nan] "not-a-number", the error message in case of non-numeric input;
-  /// defaults to 'not a number'.
+  /// [nan] "not-a-number", the error message in case of a non-numeric input;
+  /// the default value is 'not a number'.
   Num({String? nan}) : _numVal = _InputAsNum(nan, const Ok());
 
-  /// Constrains data to numeric values larger than or equal to [min].
+  /// Constrains data to numeric values greater than or equal to [min].
   ///
-  /// [min] the smallest valid input value.
-  /// [nan] "not-a-number", the error message if input is not numeric; defaults
-  /// to 'not a number'.
-  /// [small] the error message if input is too small; defaults to 'too small'.
+  /// [min] the smallest valid number.
+  /// [nan] "not-a-number", the error message if an input is not numeric; the
+  /// default value is 'not a number'.
+  /// [small] the error message if an input is too small; the default value is
+  /// 'it cannot be < [min]'.
   Num.min(num min, {String? nan, String? small})
       : _numVal = _InputAsNum(
-            nan, (num input) => input < min ? small ?? _tooSmall : null);
-
-  /// Constrains data to numeric values lesser than or equal to [max].
-  ///
-  /// [max] the largest valid value.
-  /// [nan] "not-a-number", the error message if input is not numeric; defaults
-  /// to 'not a number'.
-  /// [large] the error message if input is too large; defaults to 'too large'.
-  Num.max(num max, {String? nan, String? large})
-      : _numVal = _InputAsNum(
-            nan, (num input) => input > max ? large ?? _tooLarge : null);
-
-  /// Constrains data to numeric values within the range [min] and [max].
-  ///
-  /// [min] the smallest valid input value; must be <= [max].
-  /// [max] the largest valid input value; must be >= [min].
-  /// [nan] "not-a-number", the error message if input is not numeric; defaults
-  /// to 'not a number'.
-  /// [small] the error message if input is too small; defaults to 'too small'.
-  /// [large] the error message if input is too large; defaults to 'too large'.
-  Num.range(num min, num max, {String? nan, String? small, String? large})
-      : assert(min <= max),
-        _numVal = _InputAsNum(
           nan,
-          (num input) => input < min
-              ? small ?? _tooSmall
-              : input > max
-                  ? large ?? _tooLarge
-                  : null,
+          (num input) => input < min ? small ?? 'it cannot be < $min' : null,
         );
 
-  // default messages
-  static const _tooSmall = 'too small';
-  static const _tooLarge = 'too large';
+  /// Constrains data to numeric values that are less than or equal to [max].
+  ///
+  /// [max] the greatest valid number.
+  /// [nan] "not-a-number", the error message if an input is not numeric; the
+  /// default value is 'not a number'.
+  /// [large] the error message if an input is too large; the default value is
+  /// 'it cannot be > [max]'.
+  Num.max(num max, {String? nan, String? large})
+      : _numVal = _InputAsNum(
+          nan,
+          (num input) => input > max ? large ?? 'it cannot be > $max' : null,
+        );
+
+  /// Constrains data to numeric values within the range [min–max].
+  ///
+  /// [min] the smallest valid number; it must be < [max].
+  /// [max] the largest valid number; it must be > [min].
+  /// [nan] "not-a-number", the error message if an input is not numeric; the
+  /// default value is 'not a number'.
+  /// [small] the error message if an input is too small; the default value is
+  /// 'it cannot be < [min]'.
+  /// [large] the error message if an input is too large; the default value is
+  /// 'it cannot be > [max]'.
+  Num.range(num min, num max, {String? nan, String? small, String? large})
+      : assert(min < max),
+        _numVal = _InputAsNum(
+          nan,
+          (num input) {
+            return input < min
+                ? small ?? 'it cannot be < $min'
+                : input > max
+                    ? large ?? 'it cannot be > $max'
+                    : null;
+          },
+        );
 
   // the number validator.
   final _InputAsNum _numVal;